Added ed_file, strrep.
Issue in set,get functions for prompt and default_filename:
Buffer string was not being cleared everytime it was updated.

Eg:
	P ------>
will set the prompt to '------>'
	P ->
will set the prompt to '->---->' because the previous string in
the buffer was not cleared before setting the new string.
This is now fixed.

shopen() (wrapper to popen) added.

regerror_aux() moved to err.c

strrep() and its auxillary functions added to aux.c

Added ed_file