Erected underlying foundations for undo functions.
* Added ed_undo() and ed_redo() and its auxillary functions.

Undos and redos are unlimited and space efficient. This is achieved
with generative functions and buffers that are used by these functions.
The nature of edd (or any text editor, for that matter) is such: Any
operation can be categorized either as an "insert" or "delete". ed_
functions that insert, push their data into insert_buffer, so do
functions that delete. undoing or redoing functions look at command
they are undoing, and invert their behavior. An "append", for example,
is undone by a "delete".

The insert and delete buffers do not require any extra space - they are but
pointers to the nodes in the list. This makes them much more space
efficient than if a state-storing undo mechinism were to be used.

* ed_delete now uses ll_remove_shallow().

shallow remove only detaches a node from the global list, does not
deallocate.

* Some cleaning in ll.c, more to come.